backend/requirements.txt
# === Web Framework ===
fastapi==0.109.2
uvicorn[standard]==0.27.1

# === Database ===
sqlalchemy==2.0.27
alembic==1.13.1
aiosqlite==0.20.0
asyncpg==0.30.0

# === Validation ===
pydantic==2.6.1
pydantic-settings==2.2.1

# === File Upload ===
python-multipart==0.0.9
aiofiles==23.2.1

# === PDF Parsing ===
pdfplumber==0.10.3
pdf2image==1.17.0
PyMuPDF==1.23.26

# === OCR ===
pytesseract==0.3.10
Pillow==10.2.0

# === Data Processing ===
pandas==2.2.0
python-dateutil==2.9.0post0
numpy==1.26.4

# === Fuzzy Matching ===
rapidfuzz==3.6.1

# === HTTP / Testing ===
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.4
